requirements.txt
# PyDoll MCP Server Dependencies
# Core Requirements for Production Use

# Core MCP and Automation Dependencies
pydoll-python>=2.12.4      # Revolutionary browser automation library
mcp>=1.2.0                # Model Context Protocol core
pydantic>=2.10.4          # Data validation and settings management
pydantic-settings>=2.0.0  # Settings management with environment variable support
typing-extensions>=4.0.0  # Extended typing support for older Python versions

# Async and Performance Dependencies
asyncio-throttle>=1.0.0   # Rate limiting for requests
aiofiles>=25.1.0,<26.0.0  # Async file operations (compatible with pydoll-python 2.12.4)

# Configuration and Environment
python-dotenv>=1.0.0      # Environment variable loading

# CLI and User Interface
rich>=13.0.0              # Rich text and beautiful formatting
click>=8.1.0              # Command line interface framework

# System Dependencies
psutil>=5.9.0             # Process and system utilities for Chrome detection

# Browser and Network Dependencies (Included with pydoll-python)
# websockets>=13.1         # WebSocket client/server (via pydoll-python)
# aiohttp>=3.9.5          # Async HTTP client/server (via pydoll-python)
# beautifulsoup4>=4.12.3  # HTML parsing (via pydoll-python)

# Development Dependencies (Optional - install with pip install pydoll-mcp[dev])
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0
# pre-commit>=3.0.0
# coverage[toml]>=7.0.0

# Documentation Dependencies (Optional - install with pip install pydoll-mcp[docs])
# mkdocs>=1.5.0
# mkdocs-material>=9.0.0
# mkdocstrings[python]>=0.24.0
